--- ItspServices.pServer.Persistence/ItspServices.pServer.Persistence.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ItspServices.pServer.Repository\ItspServices.pServer.Repository.csproj" />
-    <ProjectReference Include="..\src\ItspServices.pServer.Abstraction\ItspServices.pServer.Abstraction.csproj" />
-  </ItemGroup>
-
-</Project>
