+++ src/ItspServices.pServer.Abstraction/Repository/IRepositoryPart.cs
