+++ src/ItspServices.pServer.Persistence/Repository/UserRepository.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using ItspServices.pServer.Abstraction.Models;
+using ItspServices.pServer.Abstraction.Repository;
+
+namespace ItspServices.pServer.Repository
+{
+    public class UserRepository : RepositoryPart<User>, IUserRepository
+    {
+
+        public UserRepository()
+        {
+        }
+
+        public User GetUserByName(string name)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
