+++ src/ItspServices.pServer/Startup.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using ItspServices.pServer.Persistence;
 
 namespace ItspServices.pServer
 {
@@ -32,6 +33,7 @@ namespace ItspServices.pServer
             });
 
             services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddPersistence();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
